Extract the Fedora coupled persistence methods to their own module
This will make them easier to remove when the time comes
